app/requirements.txt
fastapi~=0.112.0
uvicorn~=0.30.5
motor~=3.5.1
pydantic~=2.8.2
python-dotenv~=1.0.1